% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softbib.R
\name{softbib}
\alias{softbib}
\title{Software Bibliographies}
\usage{
softbib(
  project = getwd(),
  output = c("bib", "pdf", "docx", "rmd"),
  style = NULL,
  exclude = NULL,
  include = NULL
)
}
\arguments{
\item{project}{Path to the R project folder. By default, `softbib` uses the current working directory.}

\item{output}{Character vector with file extensions of the desired output formats.}

\item{style}{Path to a Citation Style Language file (with `.csl` extension)
which controls the reference style. CSL files can be downloaded from the
Zotero Repository: https://www.zotero.org/styles}

\item{exclude}{character vector with BibTeX keys to exclude from the software bibliography.}

\item{include}{character vector with BibTeX keys to include in the software bibliography.}
}
\description{
This function detects all the R packages used in a project folder and
automatically creates software bibliographies in PDF, Word, Rmarkdown and
BibTeX formats.
}
