% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softbib.R
\name{softbib}
\alias{softbib}
\title{Software Bibliographies}
\usage{
softbib(include = NULL, exclude = NULL, style = NULL)
}
\arguments{
\item{include}{`NULL` or character vector.
+ `NULL`: the working directory is crawled to identify `R` packages used in the project, and all packages are included in the software bibliography.
+ Character vector: the names of packages to include in the bibliography.}

\item{exclude}{character vector of package names to exclude from the bibliography.}

\item{style}{Path to a Citation Style Language file (with `.csl` extension) '
Zotero Repository: https://www.zotero.org/styles}
}
\description{
This function detects all the R packages used in a project folder and
automatically creates software bibliographies in PDF, Word, Rmarkdown and
BibTeX formats.
}
